+++ src/app/state.rs
@@ -1,0 +1,13 @@
+#[derive(Debug, Default, PartialEq)]
+pub enum AppState {
+    #[default]
+    EstablishingConnection,
+    EditingConnection,
+    Tab,
+    ReadNodeData,
+    EditCreateNodePath,
+    EditCreateNodeData,
+    EditNodeData,
+    DeleteNode,
+    ConfirmDelete,
+}
