+++ src/app_legacy.rs
@@ -86,16 +86,3 @@ impl App {
         List::from_iter(lines)
     }
 }
-#[derive(Debug, Default, PartialEq)]
-pub enum AppState {
-    #[default]
-    EstablishingConnection,
-    EditingConnection,
-    Tab,
-    ReadNodeData,
-    EditCreateNodePath,
-    EditCreateNodeData,
-    EditNodeData,
-    DeleteNode,
-    ConfirmDelete,
-}
